VsTopicRepository 수정 #93

+++ vsplay/src/main/java/com/buck/vsplay/domain/vstopic/dto/EntryDto.java
@@ -20,7 +20,7 @@ public class EntryDto {
 
     @Data
     public static class createEntriesRequest{
-        Integer topicId;
+        Long topicId;
         List<Entry> entries;
     }
 }
